ai-voice-reporting-system/backend/requirements.txt
Flask==2.1.2
google-cloud-storage==2.3.0
google-cloud-bigquery==2.34.0
google-cloud-firestore==2.3.0
google-cloud-speech==2.15.0
google-cloud-texttospeech==2.10.0
google-cloud-documentai==2.5.0
pandas==1.3.5
PyMuPDF==1.19.6
langchain==0.0.200
openai==0.27.0
flask-socketio==5.3.0
requests==2.26.0